changed int into long and remove annotation

+++ src/main/java/jkz/lab2/NumberGenerator.java
@@ -1,14 +1,14 @@
 package jkz.lab2;
 
 public class NumberGenerator {
-	private static int min = 1, max = 100;
+    private static long min = 1, max = 100;
 
-	public static void setMinMax(int min, int max) {
-		NumberGenerator.min = min;
-		NumberGenerator.max = max;
-	}
+    public static void setMinMax(long min, long max) {
+        NumberGenerator.min = min;
+        NumberGenerator.max = max;
+    }
 
-	public static synchronized int generateNumber() {
-		return (int) Math.round(Math.random() * (max - min) + min);
-	}
+    public static synchronized long generateNumber() {
+        return (long) Math.round(Math.random() * (max - min) + min);
+    }
 }
